move server test to server.rs
