--- crates/enclave-server/tests/integration/main.rs
@@ -1,6 +1,0 @@
-#[cfg(test)]
-mod server;
-
-// mod utils;
-// #[cfg(test)]
-// mod snapshot;
